+++ src/EliteChroma.Core/Layers/SrvModeSwitchesLayer.cs
@@ -1,28 +1,21 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 using Colore.Data;
-using Colore.Effects.Keyboard;
 using EliteChroma.Chroma;
 using EliteFiles.Bindings.Binds;
-using EliteFiles.Status;
 
 namespace EliteChroma.Core.Layers
 {
     [SuppressMessage("Performance", "CA1812:Avoid uninstantiated internal classes", Justification = "Instantiated by ChromaController.InitChromaEffect().")]
-    internal sealed class InterfaceModeLayer : LayerBase
+    internal sealed class SrvModeSwitchesLayer : LayerBase
     {
-        public override int Order => 1000;
-
         protected override void OnRender(ChromaCanvas canvas)
         {
-            if (Game.AtHelm && !Game.DockedOrLanded && Game.Status.GuiFocus == GuiFocus.None)
+            if (!Game.InSrv)
             {
                 return;
             }
 
-            var k = canvas.Keyboard;
-            k[Key.Escape] = Color.White;
-
-            ApplyColorToBinding(canvas.Keyboard, InterfaceMode.All, Color.White);
+            ApplyColorToBinding(canvas.Keyboard, DrivingModeSwitches.All, Color.Green);
         }
     }
 }